+++ button.lua
@@ -45,7 +45,8 @@ function GraphicsMenu()
     Buttons['renderer'] = Button(10, 190, 300, 50, function() local x = 'Screen' if NewRenderer then x = 'Canvas' end return "Renderer: "..x end, function() NewRenderer = not NewRenderer end, 0.1)
     Buttons['graphics'] = Button(10, 260, 300, 50, function() local x = 'Fast' if HighGraphics then x = 'Fancy' end return "Graphics: "..x end, function() HighGraphics = not HighGraphics end, 0.15)
     Buttons['fps'] = Button(10, 330, 300, 50, function() local x = '60' if FpsLimit ~= 71 then x = 'Unlimited'end return "FPS: "..x end, function() if FpsLimit == 71 then FpsLimit = 0 else FpsLimit = 71 end end, 0.2)
-    Buttons['back'] = Button(10, 400, 300, 50, "Back", OptionsMenu, 0.25)
+    Buttons['stepsize'] = Button(10,400,300,50,function() local x = '4' if StepSize == 10 then x = '10' end return "Step Size: "..x end, function() if StepSize == 10 then StepSize = 4 else StepSize = 10 end end, 0.25)
+    Buttons['back'] = Button(10, 470, 300, 50, "Back", OptionsMenu, 0.3)
 end
 
 function ControlsMenu()