+++ lib/playerCollision.lua
@@ -77,6 +77,14 @@ function playerCollisionDetect(tile,pB,dt) --pB formatted as "0-0, 1-0" etc.
     if blMain == 13 then
         ZoomBase = (blSub/510)+0.5
     end
+
+    --Enemy (beta)
+    if blMain == 32 then
+        Pl.animation = 'jump'
+        Pl.nextAni = 'high'
+        Pl.xv = Pl.xv * 0.9
+        Pl.yv = (0.1*Pl.yv) - 3
+    end
 end