+++ changelog.md
@@ -1,5 +1,11 @@
 Release Log:
 
+a1.0.5
+- Quarterstep adjustment setting
+- Made new Entity class
+- Hold shift to go straight into game
+
+
 a.1.0.4_01
 - Hovering over the phone now shows its hitbox
 - Phone now shrinks & grows properly